+++ scape2d/scape2d-engine/src/main/scala/scape/scape2d/engine/motion/collision/detection/detection.scala
@@ -4,7 +4,7 @@ import scape.scape2d.engine.motion._
 import scape.scape2d.engine.geom.Spherical
 
 package object detection {
-  def detectWithDiscriminant[T <: Movable with Spherical](s1:T, s2:T, timestep:Long) = {
+  def detectWithDiscriminant[T <: Movable with Spherical](s1:T, s2:T, timestep:Double) = {
     val sumOfRadii = s1.radius + s2.radius;
     val A = s1.position - s2.position;
     val B = scaleVelocity(s1.velocity, timestep) - scaleVelocity(s2.velocity, timestep);
